pyproject.toml
[tool.poetry]
name = "aoc2025"
version = "0.0.1"
description = "Solutions to Advent of Code 2025 problems"
authors = ["Daniel Halligan <dan.halligan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aocd = "^0.1"
typer = "^0.9.0"
datetime = "^5.3"
networkx = "^3.2.1"
numpy = "^1.26.2"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
aocd-example-parser = "^2023.2"
black = "^23.11.0"
pytest-cov = "^4.1.0"
ipykernel = "^6.27.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
aoc2025 = "aoc2025.cli:main"

